Add parser interface and YummyParser class. Move parse logic from controller to that class
